pyproject.toml
[project]
name = "pythonplanesappproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Karolina Jastrzębska",email = "karolina.waledzik@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "tortoise-orm[asyncpg] (>=0.24.0,<0.25.0)",
    "getter (==0.0.0.*)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
